requirements.txt
tqdm==4.64.0
gdown==4.4.0
python-dotenv==1.0.0

argparse==1.4.0


pyyaml==6.0
ruamel.yaml==0.17.21

opencv-python
numpy==1.21.0
pandas==1.3.3
seaborn==0.11.0
matplotlib==3.4.3

tensorflow==2.17.0
scikit-learn==1.0.0

jupyter==1.0.0
ipykernel==6.0.0
jupyterlab==3.4.0
